Migration of the contract to 3.11.2
